up:14 version docker db

+++ src/api/answers.py
@@ -1,7 +1,14 @@
 import uuid
 from fastapi import APIRouter
 
-from src.api.dependencies import DBDep
+from src.api.dependencies import DBDep, RoleSuperuserDep, UserIdDep
+from src.exeptions import (
+    RolesAdminException,
+    ExpiredTokenException,
+    ExpiredTokenHTTPException,
+    ObjectNotFoundException,
+    UserNotFoundException,
+)
 from src.services.answers import AnswersService
 from src.schemas.answers import AnswerAddRequest, AnswerPatch
 
@@ -11,24 +18,43 @@ router = APIRouter(prefix="/answers", tags=["Ответы"])
 @router.post("", summary="Добавление ответа")
 async def add_answers(
     data: AnswerAddRequest,
+    role_admin: RoleSuperuserDep,
     db: DBDep,
 ):
+    if not role_admin:
+        raise RolesAdminException
     await AnswersService(db).create_answers(data)
     return {"status": "Ответ добавлен"}
 
 
 @router.get("")
-async def get_answers(db: DBDep):
+async def get_answers(current_data: UserIdDep, db: DBDep):
     return await AnswersService(db).get_answers()
 
 
 @router.patch("/{answer_id}")
-async def update_answer(answer_id: uuid.UUID, data: AnswerPatch, db: DBDep):
-    await AnswersService(db).patch_answers(answer_id, data)
+async def update_answer(
+    answer_id: uuid.UUID, role_admin: RoleSuperuserDep, data: AnswerPatch, db: DBDep
+):
+    if not role_admin:
+        raise RolesAdminException
+    try:
+        await AnswersService(db).patch_answers(answer_id, data)
+    except ExpiredTokenException:
+        raise ExpiredTokenHTTPException
+    except ObjectNotFoundException:
+        raise UserNotFoundException
     return {"message": "Данные честично изменены"}
 
 
 @router.delete("/{answer_id}")
-async def delete_answer(answer_id: uuid.UUID, db: DBDep):
-    await AnswersService(db).delete_answer(answer_id)
+async def delete_answer(answer_id: uuid.UUID, role_admin: RoleSuperuserDep, db: DBDep):
+    if not role_admin:
+        raise RolesAdminException
+    try:
+        await AnswersService(db).delete_answer(answer_id)
+    except ExpiredTokenException:
+        raise ExpiredTokenHTTPException
+    except ObjectNotFoundException:
+        raise UserNotFoundException
     return {"message": "Данные удалены"}
