+++ src/api/payments.py
@@ -1,7 +1,14 @@
 import uuid
 from fastapi import APIRouter
 
-from src.api.dependencies import DBDep
+from src.api.dependencies import DBDep, RoleSuperuserDep
+from src.exeptions import (
+    RolesAdminException,
+    ExpiredTokenException,
+    ExpiredTokenHTTPException,
+    ObjectNotFoundException,
+    UserNotFoundException,
+)
 from src.schemas.payments import PaymentAddRequest, PaymentPatch
 from src.services.payments import PaymentsService
 
@@ -9,7 +16,9 @@ router = APIRouter(prefix="/payments", tags=["Платёж"])
 
 
 @router.post("", summary="Добавить платёж")
-async def create_payments(data: PaymentAddRequest, db: DBDep):
+async def create_payments(data: PaymentAddRequest, role_admin: RoleSuperuserDep, db: DBDep):
+    if not role_admin:
+        raise RolesAdminException
     await PaymentsService(db).create_payments(data)
     return {"message": "Платёж создан"}
 
@@ -20,12 +29,28 @@ async def get_payments(db: DBDep):
 
 
 @router.patch("/{payment_id}", summary="Частичное изминение данных")
-async def patch_payments(payment_id: uuid.UUID, data: PaymentPatch, db: DBDep):
-    await PaymentsService(db).patch_payments(payment_id, data)
+async def patch_payments(
+    payment_id: uuid.UUID, role_admin: RoleSuperuserDep, data: PaymentPatch, db: DBDep
+):
+    if not role_admin:
+        raise RolesAdminException
+    try:
+        await PaymentsService(db).patch_payments(payment_id, data)
+    except ExpiredTokenException:
+        raise ExpiredTokenHTTPException
+    except ObjectNotFoundException:
+        raise UserNotFoundException
     return {"message": "Данные частично изменены"}
 
 
 @router.delete("/{payment_id}", summary="Удаление данных")
-async def delete_payments(payment_id: uuid.UUID, db: DBDep):
-    await PaymentsService(db).delete_payments(payment_id)
+async def delete_payments(payment_id: uuid.UUID, role_admin: RoleSuperuserDep, db: DBDep):
+    if not role_admin:
+        raise RolesAdminException
+    try:
+        await PaymentsService(db).delete_payments(payment_id)
+    except ExpiredTokenException:
+        raise ExpiredTokenHTTPException
+    except ObjectNotFoundException:
+        raise UserNotFoundException
     return {"message": "Данные удалены"}
