+++ src/api/reports.py
@@ -1,8 +1,15 @@
 import uuid
 from fastapi import APIRouter
 
-from src.api.dependencies import DBDep
-from src.schemas.reports import ReportAddRequest
+from src.api.dependencies import DBDep, RoleSuperuserDep
+from src.exeptions import (
+    RolesAdminException,
+    ExpiredTokenException,
+    ExpiredTokenHTTPException,
+    ObjectNotFoundException,
+    UserNotFoundException,
+)
+from src.schemas.reports import ReportAddRequest, ReportPatch
 from src.services.reports import ReportsService
 
 router = APIRouter(prefix="/reports", tags=["Отчёт"])
@@ -11,8 +18,11 @@ router = APIRouter(prefix="/reports", tags=["Отчёт"])
 @router.post("", summary="Добавление отчёта")
 async def create_report(
     data: ReportAddRequest,
+    role_admin: RoleSuperuserDep,
     db: DBDep,
 ):
+    if not role_admin:
+        raise RolesAdminException
     await ReportsService(db).create_reports(data)
     return {"status": "Отчёт добавлен"}
 
@@ -23,12 +33,28 @@ async def get_report(db: DBDep):
 
 
 @router.patch("/{report_id}", summary="Частичное изминение")
-async def patch_report(report_id: uuid.UUID, data: ReportAddRequest, db: DBDep):
-    await ReportsService(db).patch_reports(report_id, data)
+async def patch_report(
+    report_id: uuid.UUID, role_admin: RoleSuperuserDep, data: ReportPatch, db: DBDep
+):
+    if not role_admin:
+        raise RolesAdminException
+    try:
+        await ReportsService(db).patch_reports(report_id, data)
+    except ExpiredTokenException:
+        raise ExpiredTokenHTTPException
+    except ObjectNotFoundException:
+        raise UserNotFoundException
     return {"message": "Данные частично изменены"}
 
 
 @router.delete("/{report_id}", summary="Удаление отчёта")
-async def delete_report(report_id: uuid.UUID, db: DBDep):
-    await ReportsService(db).delete_reports(report_id)
+async def delete_report(report_id: uuid.UUID, role_admin: RoleSuperuserDep, db: DBDep):
+    if not role_admin:
+        raise RolesAdminException
+    try:
+        await ReportsService(db).delete_reports(report_id)
+    except ExpiredTokenException:
+        raise ExpiredTokenHTTPException
+    except ObjectNotFoundException:
+        raise UserNotFoundException
     return {"message": "Отчёт удален"}
